+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Added
 
 - `Raxx.is_application?` Test to see if an application is compatable to run on Raxx Servers.
+- Typespecs for `Raxx.split_path/1` and `Raxx.set_body/2`
 
 ## [0.14.3](https://github.com/CrowdHailer/raxx/tree/0.14.3) - 2017-12-12