+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.14.4](https://github.com/CrowdHailer/raxx/tree/0.14.4) - 2017-12-28
 ## Added
 
 - `Raxx.is_application?/1` Test to see if an application is compatable to run on Raxx Servers.